+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `get_name()` to **heat pump** to read the model name.
+
 ## [1.5.0] - 2023-10-13
 
 ### Added
 
 - Added `get_system_info()` to read all system information.
-
 
 ## [1.4.3] - 2023-10-13